feat(e2e): using postgres connection pooling
