+++ api/db/dbConstants.js
@@ -1,0 +1,1 @@
+export const DB_SCHEMA = process.env.DB_SCHEMA
