+++ api/db/index.js
@@ -1,58 +1,19 @@
-import { Client } from 'pg'
+import { Pool } from 'pg'
 
-const DB_SCHEMA = process.env.DB_SCHEMA
+let pool = null
 
-const client = new Client({
-  user: process.env.DB_USER,
-  host: process.env.DB_HOST,
-  database: process.env.DB_NAME,
-  password: process.env.DB_PASSWORD,
-  port: process.env.DB_PORT,
-})
-
-export async function getTop10Records() {
-  try {
-    await client.connect()
-    const res = await client.query(
-      `SELECT data, id, "creationTimestamp" FROM "${DB_SCHEMA}".e2e_data limit 10;`
-    )
-    return res.rows
-  } catch (error) {
-    console.error(error)
-    return null
-  } finally {
-    await client.end()
+export function getPool() {
+  if (!pool || !pool.connected) {
+    pool = new Pool({
+      user: process.env.DB_USER,
+      host: process.env.DB_HOST,
+      database: process.env.DB_NAME,
+      password: process.env.DB_PASSWORD,
+      port: process.env.DB_PORT,
+      max: 5, // Maximum number of connections in the pool
+      idleTimeoutMillis: 30000, // Close idle connections after 30 seconds
+      connectionTimeoutMillis: 2000, // How long to wait for a connection from the pool
+    })
   }
-}
-
-export async function getRecordById(id) {
-  try {
-    await client.connect()
-    const res = await client.query(
-      `SELECT data, "creationTimestamp" FROM "${DB_SCHEMA}".e2e_data WHERE id = $1;`,
-      [id]
-    )
-    return res.rows
-  } catch (error) {
-    console.error(error)
-    return null
-  } finally {
-    await client.end()
-  }
-}
-
-export async function insertRecord({ data, id }) {
-  try {
-    await client.connect()
-    await client.query(
-      `INSERT INTO "${DB_SCHEMA}".e2e_data(data, id) VALUES ($1, $2);`,
-      [data, id]
-    )
-    return true
-  } catch (error) {
-    console.error(error)
-    return false
-  } finally {
-    await client.end()
-  }
+  return pool
 }
