+++ api/getLink.js
@@ -1,4 +1,5 @@
-import { getRecordById } from './db/index.js'
+import { getRecordById } from './db/getRecordById.js'
+import { getPool } from './db/index.js'
 
 export const config = { runtime: 'nodejs' }
 
@@ -11,3 +12,7 @@ export default async function handler(req, res) {
     res.status(500).json({ message: error.message || 'Internal server error' })
   }
 }
+
+process.on('SIGTERM', () => {
+  getPool().end()
+})
